Rename testConstants file
